+++ semantic-conventions/CHANGELOG.md
@@ -4,8 +4,10 @@ Please update the changelog as part of any significant pull request.
 
 ## v0.8.0
 
+- Add `name` field for events. It defaults to the `prefix`
+  ([#67](https://github.com/open-telemetry/build-tools/pull/67)).
 - Better explanation for sampling_relevant attributes
-  ([#70]https://github.com/open-telemetry/build-tools/pull/70)
+  ([#70](https://github.com/open-telemetry/build-tools/pull/70)).
 
 ## v0.7.0